Split validator's features on decorators
