+++ src/Validator.php
@@ -37,13 +37,6 @@ final class Validator implements ValidatorInterface
 
     public function __construct(
         private RuleHandlerResolverInterface $ruleHandlerResolver,
-        /**
-         * @var int What visibility levels to use when reading rules from the class specified in `$rules` argument in
-         * {@see validate()} method.
-         */
-        private int $rulesPropertyVisibility = ReflectionProperty::IS_PRIVATE
-        | ReflectionProperty::IS_PROTECTED
-        | ReflectionProperty::IS_PUBLIC,
 
         /**
          * @var bool|callable|null
@@ -62,17 +55,8 @@ final class Validator implements ValidatorInterface
      */
     public function validate(
         mixed $data,
-        iterable|object|string|null $rules = null
+        iterable $rules = null
     ): Result {
-        $data = $this->normalizeDataSet($data);
-        if ($rules === null && $data instanceof RulesProviderInterface) {
-            $rules = $data->getRules();
-        } elseif ($rules instanceof RulesProviderInterface) {
-            $rules = $rules->getRules();
-        } elseif (!$rules instanceof Traversable && !is_array($rules) && $rules !== null) {
-            $rules = (new AttributesRulesProvider($rules, $this->rulesPropertyVisibility))->getRules();
-        }
-
         $compoundResult = new Result();
         $context = new ValidationContext($this, $data);
         $results = [];
@@ -110,24 +94,6 @@ final class Validator implements ValidatorInterface
         }
 
         return $compoundResult;
-    }
-
-    #[Pure]
-    private function normalizeDataSet($data): DataSetInterface
-    {
-        if ($data instanceof DataSetInterface) {
-            return $data;
-        }
-
-        if (is_object($data)) {
-            return new ObjectDataSet($data);
-        }
-
-        if (is_array($data)) {
-            return new ArrayDataSet($data);
-        }
-
-        return new MixedDataSet($data);
     }
 
     /**
